Disable unneeded package API when not building unit tests (#10)
Disables the following package APIs when not building unit tests:

- `ITC_Id_normalise`
- `ITC_Event_normalise`
- `ITC_Event_maximise`

The goal is to reduce the core lib size (when not built for unit
testing).

This is kind of a 'compromise' solution. Ideally, these APIs should not
exist at all. However, removing them will also remove the ability to
test the functionality directly, which makes testing for edge cases
difficult.

Honestly, it's either this or adding a bunch of cumbersome and easy to
break tests, which try to indirectly test for the edge cases in the
normalisation/maximisation functionality.