Remove support for working with ID/Event subtrees (#22)
The possibility of being given an ID/Event subtree is a side effect of
having parent pointers in the ID/Event structs. This pointer is needed
to avoid resorting to recursion when traversing the trees.

The ITC specification (as presented in the research paper) does not use
parent pointers and thus does not know whether a given ID/Event tree is
a subtree of another node. It just treats all nodes as if they are
roots.

For this reason, it is at the very least pointless to allow subtrees to
be passed in as input to the public API, as they have no
special/different meaning in the ITC specification (in fact they don't
even exist there). Passing a subtree node as input is likely a sign that
something is wrong with the ID/Event node itself. Such a node should not
be used as even though the code _should_ handle it correctly (treating
it as if it were a root node), it may still lead to undefined and
undesired behaviour since the caller most likely passed the subtree node
by mistake and instead wanted to pass in a root node.